fix off-by-one error in sent_split [fixes #7]
The sent_split function wants to check that if there is more than one
part to a line after splitting on sentence punctuation and then if the
line's last part doesn't end in a sentence splitting punctuation, that
last part will need to be joined with the following line.  The issue was
that the right way to get all but the last part is parts[0:-1], not
parts[0:-2].  The latter simply creates the empty list when parts is
only of length 2.